Added OptionsMenu to Game and SFX for attack and shield hits
Options Menu can be open and closed using button on top right. Music volume can be adjusted the same way as in Menu scene, and there's the option to restart the game from level 1 and return to the main menu. Also added sfx for attack and shield hits to the audio manager so you can get them to play.